add places on the map

+++ app/src/main/java/com/cheerz/casadelrestau/network/data/MiamzLogin.kt
@@ -10,8 +10,8 @@ data class MiamzReqSignUpUser(val email: String, val password: String)
 
 data class MiamzReqPlaceData(
         val name: String,
-        val lng: Float,
-        val lat: Float,
+        val lng: Double,
+        val lat: Double,
         val first_image_url: String,
         val second_image_url: String,
         val third_image_url: String,
@@ -23,4 +23,4 @@ data class MiamzReqPlaceData(
 
 data class MiamzReqPlace(val place: MiamzReqPlaceData)
 
-data class MiamzPlace(val lat: Float, val lng: Float, val meter_distance: Int)
+data class MiamzPlace(val lat: Double, val lng: Double, val meter_distance: Int)
